refactor: 그룹 컨트롤러를 그룹 Find, Manage 컨트롤러 두개로 분리
