+++ backend/rush/src/main/java/rush/rush/controller/FindGroupController.java
@@ -1,17 +1,12 @@
 package rush.rush.controller;
 
-import java.net.URI;
 import java.util.List;
 import lombok.RequiredArgsConstructor;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.PathVariable;
-import org.springframework.web.bind.annotation.PostMapping;
-import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RequestMapping;
-import org.springframework.web.bind.annotation.RequestParam;
 import org.springframework.web.bind.annotation.RestController;
-import rush.rush.dto.CreateGroupRequest;
 import rush.rush.dto.GroupResponse;
 import rush.rush.dto.GroupSummaryResponse;
 import rush.rush.dto.SimpleUserResponse;
@@ -22,32 +17,16 @@ import rush.rush.service.GroupService;
 @RequiredArgsConstructor
 @RestController
 @RequestMapping("/groups")
-public class GroupController {
+public class FindGroupController {
 
     private final GroupService groupService;
-
-    @PostMapping
-    public ResponseEntity<Void> create(@RequestBody CreateGroupRequest createGroupRequest, @CurrentUser UserPrincipal userPrincipal) {
-        Long groupId = groupService.createGroup(createGroupRequest, userPrincipal.getUser());
-
-        return ResponseEntity.created(URI.create("/api/groups/" + groupId))
-                .build();
-    }
-
-    @PostMapping("/join")
-    public ResponseEntity<Void> join(@RequestParam(value = "invitation_code") String invitationCode, @CurrentUser UserPrincipal userPrincipal) {
-        Long groupId = groupService.join(invitationCode, userPrincipal.getUser());
-
-        return ResponseEntity.created(URI.create("/api/groups/" + groupId))
-                .build();
-    }
 
     @GetMapping("/mine")
     public ResponseEntity<List<GroupSummaryResponse>> findMyGroups(@CurrentUser UserPrincipal userPrincipal) {
         List<GroupSummaryResponse> groupSummaryResponses = groupService.findAllByUser(userPrincipal.getUser());
 
         return ResponseEntity.ok()
-                .body(groupSummaryResponses);
+            .body(groupSummaryResponses);
     }
 
     @GetMapping("/{id}")
@@ -55,7 +34,7 @@ public class GroupController {
         GroupResponse groupResponse = groupService.findOne(groupId, userPrincipal.getUser());
 
         return ResponseEntity.ok()
-                .body(groupResponse);
+            .body(groupResponse);
     }
 
     @GetMapping("/{id}/members")
